Fix bug in simulator
This fixes a bug where a single mine takes as much resources from a deposit as available
(which may be more than three units if multiple mines are connected to the deposit) instead
of a maximum of three resources.

If the multiple mines connected to the deposit have paths of different length to the common
factory, this results in a wrong result.