Prevent violating time limit
It turns out the given 'time to accumulate' was too short in some cases,
especially when the overall runtime limit was low (< 2s).
Now, the time for accumulation (and printing the result!) is clamped between
500 and 1500 milliseconds, which seem to be enoug, as measurement with
different runtimes and tasks lead to the conclusion that accumulating and
printing the result takes around 300ms.